desde la compu de saul. todo revisar la clase que no sirvio

+++ src/main/java/com/grid/inventorymanager/repository/AssetMovementsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssetMovementsRepository extends JpaRepository<AssetMovements, Integer> {
+public interface AssetMovementsRepository extends JpaRepository<AssetMovements, Long> {
 }
